scale double

+++ src/lightbeam.cc
@@ -1,7 +1,7 @@
 #include "lightbeam.h"
 
-ElectronBeam::ElectronBeam(int pos, float scale)
+ElectronBeam::ElectronBeam(int pos, double scale)
 {
   position = std::make_unique<int>(pos);
-  xscale = std::make_unique<float>(scale);
+  xscale = std::make_unique<double>(scale);
 }
